np.float deprecated, use float

+++ tests/test_convert/test_enrich2/test_utilities.py
@@ -73,7 +73,7 @@ class TestIsNumeric(unittest.TestCase):
         self.assertTrue(utilities.is_numeric(int))
 
     def test_true_for_np_float(self):
-        self.assertTrue(utilities.is_numeric(np.float))
+        self.assertTrue(utilities.is_numeric(float))
 
     def test_true_for_np_int(self):
         self.assertTrue(utilities.is_numeric(np.int))
